requirements.txt
aiogram~=3.24.0
python-dotenv~=1.1.1
asyncpg